refactor(web): декомпонувати chatActions/types.ts 672 LOC на 6 доменних шарів (initiative 0001)
Розбиваю apps/web/src/core/lib/chatActions/types.ts (672 LOC) на 6
доменних файлів-сусідів:

- types.result.ts (34 LOC) — ChatActionResult + ChatActionUndoableResult.
- types.finyk.ts (188 LOC) — finyk action payload-и + Budget/Debt/
  Receivable/MonthlyPlan domain entities.
- types.fizruk.ts (135 LOC) — fizruk action payload-и + WorkoutSet/
  WorkoutItem/Workout entities.
- types.routine.ts (85 LOC) — routine/habit action payload-и +
  HabitState entity.
- types.nutrition.ts (103 LOC) — nutrition action payload-и +
  NutritionMeal/NutritionDay entities.
- types.cross.ts (116 LOC) — крос-модульні (briefing, summary,
  goal, аналітика, notes, memory, утиліти) + CompareWeeksModule.

types.ts (281 LOC) — barrel: re-export усього з доменних файлів і
дискримінованого union-у ChatAction (бо handler dispatch і Anthropic
tool-schema спираються саме на нього).

Жоден consumer не змінює свої import-и: серверAction-и, finyk/fizruk/
routine/nutrition/crossActions, hubChatActionCards, hubChatActions усі
далі резолвлять './types' / '../types'.

eslint.config.js: знімаю файл з max-lines allowlist — нові файли
≤ 281 LOC, max-lines: 600 тримає себе сам.

Phase 2 PR initiative-0001-module-decomposition.md, файл #3 з 5.
